+++ go/main.go
@@ -10,7 +10,7 @@ import (
 	"os"
 )
 
-const Version = "2.4.0"
+const Version = "2.4.1"
 
 const helpText = `srv - run commands on a remote SSH server with persistent cwd.